// src/app/lib/enums.ts
export enum FileType {
  Any = 1,
  Image = 2,
  Audio = 3,
  Video = 4,
  Excel = 5,
  Word = 6,
  Pdf = 7,
  Specific = 8,
}

export enum WeekDay {
  Sunday = 0,
  Monday = 1,
  Tuesday = 2,
  Wednesday = 3,
  Thursday = 4,
  Friday = 5,
  Saturday = 6,
}

export enum FormElementType {
  Button = 0,
  Input = 1,
  Captcha = 2,
  AutoComplete = 3,
  Checkbox = 4,
  ColorPicker = 5,
  DatePicker = 7,
  DropDown = 8,
  Editor = 9,
  File = 11,
  Radio = 13,
  Switch = 14,
  Tag = 15,
  Label = 18,
  Number = 19,
  TimePicker = 20,
  PhoneNumber = 21,
  MultiSelect = 22,
}

export enum DropdownKnownList {
  Countries = 1,
  Zones = 2,
}